+++ backend/ModelLayer/DataTransferObjects/ContactPossibilitiesEntryDto.cs
@@ -1,5 +1,7 @@
-﻿using System;
+﻿using ModelLayer.Helper;
+using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Text;
 
 namespace ModelLayer.DataTransferObjects
@@ -13,7 +15,10 @@ namespace ModelLayer.DataTransferObjects
 
     public class ContactPossibilitiesEntryCreateDto
     {
+        [Required(AllowEmptyStrings = false)]
         public string ContactEntryName { get; set; }
+
+        [MailAndPhoneValidator]
         public string ContactEntryValue { get; set; }
     }
 }
